hpcviewer: Limit the number of characters in the title of a graph
The title of the graph is based on the name of the clicked scope in the
tree.
For C++ functions/templates, the name of the scope can ridiculously long
and doesn't fit the editor. It's important to limit the number
characters for the sake of visibility.

ApplicationWorkbenchAdvisor.java:
Add listener only during opening window event instead of during
activated window event. The former is triggered once, while the later is
triggered every time the user activate the viewer

BaseScopeViewActions.java: Removed unused imports

Preparation for the new release: Rename the release to 2019.07