hpctraceviewer: fixed issue #48 (typing depth manually)
Eclipse SWT doesn't allow spinner to be higher than the maximum value
(which is the maximum depth of call stack).

If a user type incorrect value or value higher than the maximum depth,
then we just display an error message. It seems SWT doesn't allow to
change the text of the spinner (at least there is no API for that).

This commit fixed issue #48