hpcviewer: performance bug on Windows (setting column's width)
On Windows, it takes time to set the width of a column although the
redraw is disabled. 
Setting columns' width is needed on Linux GTK to force table columns to
be adjusted according to the default width. Somehow on Eclipse Neon this
doesn't work properly using table layout.

The solution is to force setting column width only on Linux platform.