hpcviewer: fix issue derived metrics in thread view
Thread view is different than other views (top-down, bottom-up and flat)
that it doesn't rely on metrics in the experiment. Instead, it relies on
metrics in metric-db (also called "raw metrics).

When we create a derived metric inside thread view, we don't want to
propagate to "normal" views as the metrics are different.

A derived metric created inside Thread view should be isolated within
the view.