hpcviewer: fix issue 51 (unsuppressed empty columns)
Problem: dynamic views don't have information which columns are
hide/show because the tree is not built yet, and the columns are not
created. 

Solution: move hide/show metric computation in hpcdata instead in
hpcviewer. 
We lose the original information of hide/show metrics from
experiment.xml, but it doesn't matter since we want to hide at the start
of hpcviewer.