Sparse column: Only create columns if the metric of the root has value
One of the cause of scalability issues is too many columns are created
even if they are hidden due to empty metric value.
On Windows, every time we resize a column, the OS notifies all other
columns to resize adjust their position (including the width). Somehow,
this causes very long time to compute the new position.

This patch will NOT create columns for empty metric. However, when users
displays hide/show columns, it still shows the empty metrics on the
list. This is just to tell users that the metrics are there, but the
columns can't be shown.