hpcviewer: framework for sparse metric columns
Ensure "show column properties" functionality to be able to handle
sparse metric columns.
In the beginning, hpcviewer assumed that the number of metrics is the
same as the number of columns. For sparse metric columns, this is not
the case. 
Hence, this patch ensures that"show column properties" function to
hide/show columns, can work well with sparse metric columns.