hpctraceviewer: add a feature to show the procedure name(s)
New features (some are merged from release-2019.11):

- add capability to associate between RGB code and procedure name. Due
to limitation on Windows, an RGB code can be associated with multiple
procedures :-(
- add generic tooltip to show the procedure name of the current pixel.
  This tooltip shows the procedure name on most views (main, depth,
summary).
- add a new view to show the list of procedures and its percent. Instead
of showing a pie chart, we show a table of procedures. The reason is
that there is no simple and light way to create a pie chart in Eclipse.
Eclipse BIRT chart is too heavy weight, not suitable for us. While
SWTChart only supports bar, line and scatter charts.