hpctraceviewer: fix issue #56 (statistic view)
Issue #56 only happens on Linux, and doesn't occur on Mac. 
Somehow summary view doesn't send the data to statistic view for the
first time. Probably because on Linux, summary view is painted early and
statistic view is not created (or not registered) yet. Hence, the
message from the former doesn't reach the later.

The solution is to force summary view to broadcast the data when
statistic view requests it.

Minor update:
- add a map for "reserved colors".
- rename ??? to [No activity]