hpctraceviewer: more handling in cancel request
Every time we resize the window, it notifies main canvas to refresh the
data and schedule a "job" to collect and paint new data.
However, if there are already jobs in waiting or in execution, we have
to cancel them to ensure only one job is doing the work.

Some time, a job can't be canceled because some of its threads are still
working and don't get cancellation notice. We need to spread the
checking of the cancellation state to ensure all threads are aware if
the work needs to be canceled.

This patch improves reliability of hpctraceviewer. However, on Linux
VirtualBox (with one processor), we still encounter deadlock.

Note: a job can be a group of Java threads.