hpcviewer: fix metric columns when applying filter
Symptom: Applying filter in hpcviewer causes empty metric columns. 

Cause: Applying filter forces hpcviewer to recreate tree and the metrics
as well.
Hence, the metric reference in the table is no longer the same as the
metric reference in the experiment class. 

Solution: not using object reference when comparing metrics. Should compare with metric index or the name.