Fix derived metric which requires the value of a "future" metric
Some times, hpcrun derived metrics require the value of "future"
metrics. This causes the value of derived metrics to be empty. 

Solution: compute derived metrics at the end. Hence, we are more
guaranteed that the required value is available.

Example:
    <Metric i="652" n="GKER:STMEM (B):Sum (I)" o="46" >
      <MetricFormula t="view" frm="#38/#54"/>
    </Metric>

In this case, this metric (order #46) requires the value of a "future"
metric (order #54) which is not available yet.