Fix infinite recursive bug when all GPU metrics are zero
Problem: Some GPU metrics contain derived metrics which depend on the
value of other metrics.
The problem is hpcviewer assumes the value of required metrics are
always available. If the value is not available, the viewer tries to
compute recursively the value.

Solution: add condition to avoid a metric to find its own derived value.

Minor update: clean-up the code, and add the metric's name in the
tooltip so users know if the metric is inclusive or exclusive (if the
column header is cropped)