hpcviewer: fixed issue #69 (View menu disappeared)
The bug occurs when we open 2 hpcviewer windows: one with a database,
the other one (the last one) is with no database. As Eclipse only allows
one menu-variable (whether the menu should be visible or not) per
application for all windows, it hides "View" menu for all hpcviewer
windows.

Solution: instead of using Eclipse declarative menu, we always 
enables View menu, but programmatically detects if we should execute the
command or not. For instance, before showing Metric dialog, we need to
check if in this window, there are opened databases or not.