Partial Fix issue #71 (stripped hpcstruct confuses hpcviewer)
This patch fixes the original issue #71 when hpcviewer sometimes
crashes, sometimes displays blank rows (depending on the Eclipse's
version) when displaying the bottom-up view of stripped hpcstruct
database.

The root cause of the problem is because Eclipse, under the hood, uses
the object's hashcode to identify the path of a tree item. Since we
override the hashcode() method by returning the flat index of the node,
it confuses Eclipse SWT to uniquely identify a tree path (since there
are more than one identical paths). 

The hashcode() code we wrote in Scope class is dated 20 years ago (or
perhaps more?) when hpcviewer is simple and written in pure Java. It
worked nicely for simple database, but broken in Eclipse SWT for weird
database like stripped hpcstruct.

Solution: Uses Java's System.identityHashCode to compute the node's hashcode instead of using flat or cct index.