Partial fix issue #71 (bottom-up tree on stripped hpcstruct)
Allow combining metrics when the target and the source node are at the
same level in CCT.
If we have a CCT:
 root -> b -> a1 -> a3
 root -> b -> a2 -> a4
 
Then when creating bottom-up tree dynamically, we should allow combining
metrics a1 to a2.

Note: This patch works in some cases, but it doesn't work in other cases.