hpcviewer: Fix columns move by themselves
Perhaps this is an Eclipse bug with tables using lazy virtual content
provider.
Once the table content is initialized and Eclipse SWT finalizes loading
the content, the position and the size of the table columns can be
incorrect. Worse, when users resize the table, the columns move by
themselves trying to adjust their position and size.

To avoid this behavior, we have to force Eclipse to refresh the table
once we set the input. Ideally, we refresh once Eclipse finishes loading
the content, but unfortunately there is no callback for this event.
Refreshing after setting the input is not ideal, but it works in most
cases.