hpctraceviewer: Fix deadlock on Mac OS
This bug only happens on Mac exported version, but runs well under
Eclipse and other platforms.

It seems the root cause of the deadlock is when resizing the window or
change the depth or the cursor position very fast that more than 1
Eclipse jobs are in the pool.

Solution: avoid to submit multiple jobs. 

If there is an Eclipse job currently running, the new job needs to send
request cancel it, and then submit a new job. Eclipse doesn't guarantee
the job will be canceled, so we need to be careful.

Caveat: this approach causes deadlock in VirtualBox Ubuntu with one core (but works fine on VirtualBox Windows).