hpctraceviewer: fix issue #75 (deadlock on Mac OS)
Root cause: when call stack view is initialized or updated, it resets
the depth to zero. This triggers a notification to other views to
refresh their data, including the main view. The later which just starts
scheduling a job, has to request to cancel the current job. 
A rendering job consists of two thread groups: one for collecting data
(producer) and the other to paint to canvas (consumer). It seems the
producer receives cancellation, but not the consumer. The consumer then
waits forever data from producer which has been terminated. 

Solution: do not reset the depth in call stack view. Let the main view set the depth.