hpcviewer: avoid creating bottom-up and flat tree when exiting the app
Sometimes Eclipse fires activate event to some views (especially dynamic
views) when the application exits.
It seems the root cause of this is when we the application closes (or
one of the windows closes) it sends "hide" event to a flat view, which
causes the bottom-up view to be "visible" and then "active".

This bug is not reproducible easily, and it isn't clear why.

The "solution" is not to hide the views when a window closes, but we
hide the views in the beginning of the creation.

Also, we need to associate a workbench window with a window state. 
If the window's state is closing, we shouldn't create dynamic trees.