hpctraceviewer: add the time axis in the main view
Add time label in the main view so that users can see time coordinates
easily (instead of forcing users to click to know the time).

The time axis can change adaptively according to the current range: if
the range changes, the axis also changes. 
However, at the moment we cannot change the time range from the axis.