hpcviewer: fix last column for Linux gtk ppcle64 platform
This is specific to Linux gtk ppcle platform where SWT doesn't want to
show again the hidden last column.
We have to force to set the layout to trick SWT to show the column.

Note:
Code inside SWT's TreeColumnLayout.class shows that layout is set if the
platform isn't gtk and isn't the last column. 

		if( ! IS_GTK || t.getColumn(t.getColumnCount()-1) != tColumn ){
			tColumn.setData(LAYOUT_DATA,new ColumnPixelData(tColumn.getWidth()));
			layout(t.getParent(), true);
		}