hpcviewer: check Java version via Java properties
Checking Java version from command line is error prone and it highly
depends on the output of each Java programs. There is no standardized
Java output, hence it's too tricky to parse the command line.
A more reliable way is to use Java system properties, which is more
"standardized" than the command line.

For the error message, since we cannot use any Eclipse SWT API during
the launch time, we have to mix the code with Java Swing to display an
error message.