Add a feature to check X display inside Java.
The viewer launch script already checks if the display variable is
properly set or not. However, it happens users don't run the viewer from
launch script, and it fails with Java segfault (and hard to decode).

Fix: Add a new feature inside Java to check the display. This reduces
the number of SWT error messages (although it doesn't eliminate them
completely).

The trickiest part is to put where in the program we should detect the
display. If it's too early, and we exit too early Eclipse/JVM will throw
error messages. If it's a bit late, we'll encounter SWT error (since the
viewer setups splash screen very early). 
At the moment there's no ideal solution. But it seems detecting inside
IApplication instance is better since it generates less error messages

Minor update: restructure Java version code.