Fix issue #80 (last column cannot be stretched)
The problem is that we use relative width on the first column which
causes SWT to recompute the width every time a user changes the width of
the last column.

The fix is to use absolute width. However, this causes the width of the
first column becomes not-proportional to metric columns.

At the moment it's better to look disproportional.