hpcdata: Remove invisible procedures from the tree
hpcprof (in fix-prof_viewer branch) provides an additional feature to
specify if a node (or a procedure) has to be invisible or not.
Sometimes we want to hide nodes from hpcviewer (like <no activity> node)
but we want it to appear in hpctraceviewer.

This commit allows hpcdata (and hpcviewer) to parse new xml from hpcprof
to mark which nodes have to be hidden in the viewer.

See hpctoolkit src/lib/prof/NameMapping.cpp for further details.