hpcviewer: add an API to return a list of "visible" metrics
Add a new API: getVisibleMetrics()

Some metrics like GPU*ACCU shouldn't be displayed and visible for end
users. We use them to compute derived metrics and aggregate metrics to
create bottom-up and flat trees dynamically.

We keep the old API getMetrics() to retrieve the complete list of
metrics, including "helper" metrics that are not visible to users.