hpcviewer: fix issue #86 (contexts in bottom-up tree)
The bottom-up tree doesn't include the context when creating dynamically
the tree. It merges all call chains that share the same procedure.
However, this can confuse users as it's hard to correspond to the
top-down tree.

One solution is we merge the call chains if they share the same
procedure AND the line number of the call sites is the same. 

Warning:
This solution works in most cases. If the executable has weird macros
(like amg2006), hpcstruct may generate different struct IDs for
different instances of macros (or procedure frame). In this case, some
call chains cannot be merged although they have the same macro name and
the same line number.