Fix metric font by not using Java AWT font names
Mixing Java AWT and Eclipse SWT is really a bad idea. The name of the
font in Java AWT is not the same as the one in Eclipse SWT. 
When Java AWT returns the list of available fonts for the current OS,
some of those fonts don't work well with Eclipse SWT (and unfortunately
SWT doesn't complain). So we never know if a font can work with SWT or
not.

A better way is to use JFace font to get fixed font names. So far it
looks working fine on Mac and Linux.