Fix issue #34 (sorting a metric column raises an exception)
This corner case occurs for a very particular database with a specific
steps to reproduce (see issue #34).
The problem happens if both values are empty and both nodes have exactly
the same name with inconsistent line number.

Solution: If the metrics are the same, use the name comparison as the second sorting criteria.